+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/userinfo/Contact.kt
@@ -5,5 +5,7 @@ import androidx.annotation.Keep
 /** Contact object for miniapp. */
 @Keep
 data class Contact(
-    val id: String
+    val id: String,
+    var name: String?,
+    var email: String?
 )