+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/userdata/ContactListActivity.kt
@@ -3,7 +3,6 @@ package com.rakuten.tech.mobile.testapp.ui.userdata
 import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Context
-import android.content.DialogInterface
 import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
@@ -17,8 +16,11 @@ import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.rakuten.tech.mobile.miniapp.js.userinfo.Contact
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.ContactsActivityBinding
+import com.rakuten.tech.mobile.testapp.helper.defaultContact
+import com.rakuten.tech.mobile.testapp.helper.isEmailValid
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import java.util.UUID
@@ -67,30 +69,49 @@ class ContactListActivity : BaseActivity() {
     }
 
     private fun onSaveAction() {
-        settings.contactNames = adapter.provideContactEntries()
+        settings.contacts = adapter.provideContactEntries()
         finish()
     }
 
     private fun onAddAction() {
         val contactView = layoutInflater.inflate(R.layout.dialog_add_contact, null)
-        val editNewContact: AppCompatEditText = contactView.findViewById(R.id.editNewContact)
+        val edtContactId = contactView.findViewById<AppCompatEditText>(R.id.edtContactId)
+        val edtContactName = contactView.findViewById<AppCompatEditText>(R.id.edtContactName)
+        val edtContactEmail = contactView.findViewById<AppCompatEditText>(R.id.edtContactEmail)
+
         ContactInputDialog.Builder().build(this).apply {
             setView(contactView)
-            setPositiveListener(DialogInterface.OnClickListener { dialog, _ ->
-                if (editNewContact.text.toString().isNotEmpty()) {
-                    val name = editNewContact.text.toString()
-                    // add contact in the last position
-                    adapter.addContact(adapter.itemCount, name.trimEnd())
+            setPositiveListener(View.OnClickListener {
+                var canSave = true
+                val id: String = edtContactId.text.toString().trim()
+                var name: String? = edtContactName.text.toString().trim()
+                var email: String? = edtContactEmail.text.toString().trim()
+
+                if (id.isNotEmpty()) {
+                    if (name!!.isEmpty())
+                        name = null
+
+                    if (email!!.isEmpty())
+                        email = null
+                    else {
+                        if (!email.isEmailValid()) {
+                            canSave = false
+                            Toast.makeText(this@ContactListActivity, getString(R.string.userdata_error_invalid_contact_email), Toast.LENGTH_LONG)
+                                .apply { setGravity(Gravity.TOP, 0, 50) }
+                                .show()
+                        }
+                    }
                 } else {
-                    Toast.makeText(
-                        this@ContactListActivity,
-                        getString(R.string.userdata_error_invalid_contact),
-                        Toast.LENGTH_LONG
-                    ).apply {
-                        setGravity(Gravity.BOTTOM, 0, 100)
-                    }.show()
+                    canSave = false
+                    Toast.makeText(this@ContactListActivity, getString(R.string.userdata_error_invalid_contact_id), Toast.LENGTH_LONG)
+                        .apply { setGravity(Gravity.TOP, 0, 50) }
+                        .show()
                 }
-                dialog.cancel()
+
+                if (canSave) {
+                    adapter.addContact(adapter.itemCount, Contact(id = id, name = name, email = email))
+                    this.dialog?.cancel()
+                }
             })
         }.show()
     }
@@ -102,24 +123,21 @@ class ContactListActivity : BaseActivity() {
 
     private fun renderContactList() {
         if (!isFirstLaunch && settings.isContactsSaved) {
-            if (settings.contactNames.isEmpty()) {
+            if (settings.contacts.isEmpty()) {
                 renderAdapter(arrayListOf())
-            } else renderAdapter(settings.contactNames)
+            } else renderAdapter(settings.contacts)
         } else {
-            val randomList = createRandomUUIDList()
+            val randomList = createRandomContactList()
             renderAdapter(randomList)
         }
     }
 
-    private fun createRandomUUIDList(): ArrayList<String> {
-        val randomUUIDs = ArrayList<String>()
+    private fun createRandomContactList(): ArrayList<Contact> = ArrayList<Contact>().apply {
         for (i in 1..10)
-            randomUUIDs.add(UUID.randomUUID().toString().trimEnd())
-
-        return randomUUIDs
+            this.add(defaultContact(UUID.randomUUID().toString().trimEnd()))
     }
 
-    private fun renderAdapter(contactNames: ArrayList<String>) {
+    private fun renderAdapter(contactNames: ArrayList<Contact>) {
         adapter.addContactList(contactNames)
         binding.listContacts.adapter = adapter
         binding.listContacts.layoutManager = LinearLayoutManager(applicationContext)